Changed Registration so that Hospital Insurance types get linked to a new Hospital Firm where the insurance number is set to the patient id
Changed the controllers file so that the program checks for an insurance 
firm called "Hospital" and adds it if it doesn't exist.  If it exists, 
then it sets the program variable HospitalFirmId to this insurance firm.

Changed the model_inventory so that customer totals now include 
insurance values.  Previously, I was only counting customer payments, 
but now, I go through and add in any insurance payments so that the 
totals on the right hand side of billing add up.